Implement assertException for PHPUnit
Problem:
PHPUnit asserts exceptions via the expectException methods. The problem with this approach is that it allows the exception to propogate out of the unit test method, preventing any kind of valication after the fact. Since I'm trying to write London style TDD, I need to verify that specific methods were called a certain number of times with specific arguments, even if the method being tested ended up throwing an exception.

Fix:
Instead of expecting an exception, I wrote a method that takes an expected exception and a closure. The method will call the closure and assert that the expected exception was thrown. If it was, it will allow the test to continue and allow further verification to take place. If either no exception is thrown, or an exception with the wrong class or message is thrown, the assertion will fail the test.