Only return recent noun for initial key segment
\## Problem
The intended use case of the noun store is to keep track of recently referenced nouns, that is why the most recent noun is returned. So if you continually reference and store users in the noun store, when you make a statement like "And the User can access the secure portal", the noun "User" is referring to the most recent user that you referenced, just as would be the case in normal English language.

A problem arose when this logic was extended to the recently implemented complex noun navigation, where if any noun in the key was array like, and no nth was specified, then the most recent entry would be returned.

\### Simple key example (one segment):
If you asked for "Person", and there were three people stored under that key, you would get the most recent person, rather than the entire "Person" collection.

\### Complex key example (multiple segments):
If you asked for "Person's car", and "car" was a collection of cars, you would get the most recent car, rather than the entire "car" collection.

This caused a problem with Laravel models because they *are* array-like. So when you wanted to a relationship from the store, for example, a customers address, and the Customer Model had a relationship to an Address model, the noun store would see that you asked for "Customer's address", fetch the "address" relationship, see that it was array-like and return the "end" of it instead of the entire object.

Given the intended use case of the noun store, it really makes no sense for any noun other than the first entry in the key to return the "end", so I have modified the behavior to do just that:

\## Modified behavior:

\### Simple key example (one segment):
If you asked for "Person", and there were three people stored under that key, you would get the most recent person, rather than the entire "Person" collection (because "Person" is the first part of the key).

\### Complex key example (multiple segments):
If you asked for "Person's car", and car was a collection of cars, you will get the entire collection of cars. Since "car" is not the first segment of the key, the noun store will not return the "end" of it.